devops: configure db service in User docker-compose file
Introduce a Dockerfile for building a custom postgres image
that supports custom server config, authentication config and
configuration of conventional PostgreSQL environment variables
via an .env file. See `.env.template` for more information.

Include a step in the db Dockerfile to install wal2json
to support JSON output for replication slot. The purpose is
to enable the implementation of Transactional Outbox pattern
down the line.